ResourceTree reacts to deletion of tags while not visible
If inlineTree is false, the tree is only shown in a popup
and up to now did not reflect changes which were made
in the choices (deletion of one by clicking on the delete
button of the tag itself).

By adding the tree to the DOM (and hiding it), it receives
events and is kept up to date with the content of the
tags choices.